Added groundwork to implement an update checker in the future, as of right now the small "update available" symbol only shows that a mod has an updateJson link present, not that there is an update available.
Added the ability to sort mods in the list based on which one has an available update. (at the same time reworked how mods are searched)

Attempted to fix problem where some mods with a blank but not null icon file would not show their icon in the list view.
Changed the background of the mod content page from blank to white