The program is now able to check if there are updates available for a given mod. Instead of doing this for all mods at once using a separate function, each mod is responsible for checking if it has an update itself which is done in the background. If a modinfo file does not have a updateJson url then this is skipped.
Mod update version is displayed in a tooltip when the mouse hovers over the diamond (shown on mods which have updates available).

Slightly increased the size of the diamond shown on mods with updates to make hovering over with a mouse easier.

Fixed a problem where mods with the examplemod icon set would show with no icon, it now correctly shows the fallback icon. (Mods should always have an icon now).

Slightly increased the minimum width of the mod listview to fix some clipping issues.

Removed some uneccesary classes (logger will be used in the future so it can stay).

Cleaned up some code.